requirements.txt
asgiref
certifi==2022.12.7
chardet==4.0.0
Django==4.1.7
djangoajaxswiss==4.1
gunicorn==20.1.0
idna==3.3
jsonfield==3.1.0
openpyxl==3.0.9
Pillow==9.3.0
psycopg2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3
uuid
boto3==1.21.3
django-storages==1.12.3
django-picklefield~=3.1.0
faker
selenium
factory-boy
faker_education
faker_biology